requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
python-jose[cryptography]
psycopg2-binary==2.9.9
bcrypt==4.0.1
pendulum==3.0.0
